api-gateway/requirements.txt
# Core Framework
# FastAPI 0.115.x+ fully compatible with Pydantic v2 (2025)
fastapi==0.115.7
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.11.0

# Database
# SQLAlchemy 2.0.44 latest stable with full asyncpg support
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.14.0

# Auth & Security
# PyJWT recommended by FastAPI (replaced python-jose in 2025)
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
email-validator==2.2.0

# Task Queue
# Celery 5.5.x with Redis compatibility (redis<=5.2.1 for kombu)
celery==5.5.0
redis==5.2.1

# HTTP Client
httpx==0.28.1

# DOCX Processing
python-docx==1.1.2
Pillow==11.0.0

# Template Engine (S2-04)
jinja2==3.1.5

# Testing
# pytest 8.4.2 supports Python 3.9-3.14
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==6.0.0

# Code Quality
ruff==0.9.1
black==24.10.0
mypy==1.14.1

# Utilities
python-dotenv==1.0.1
